fix: rename LLMClient to llm_client for PEP 8 compliance
- Rename LLMClient directory to lowercase llm_client
- Update all import statements throughout codebase
- Maintain Git history through proper git mv operations
